+++ lib/consul_application_settings/settings_provider.rb
@@ -2,8 +2,8 @@ module ConsulApplicationSettings
   # Provides access to settings stored in Consul or in file system
   class SettingsProvider
     def initialize(base_path, config)
-      @consul_provider = ConsulProvider.new(base_path, config)
-      @file_provider = FileProvider.new(base_path, config)
+      @consul_provider = ConsulApplicationSettings::Providers::Consul.new(base_path, config)
+      @file_provider = ConsulApplicationSettings::Providers::LocalStorage.new(base_path, config)
     end
 
     def get(path)
